backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Api ShopEasy"
requires-python = ">=3.13"
dependencies = [
    "bandit>=1.7.0",
    "black==25.1.0",
    "certifi==2022.6.15",
    "chardet==3.0.4",
    "coreapi==2.3.3",
    "coreschema==0.0.4",
    "django==5.2.6",
    "django-filter==25.2",
    "django-cors-headers==4.1.0",
    "django-debug-toolbar>=4.4.5",
    "django-minio-storage>=0.5.8",
    "django-rest-swagger==2.2.0",
    "django-templated-mail==1.1.1",
    "djangorestframework>=3.9.3",
    "djangorestframework-simplejwt>=4.4.0",
    "djoser==2.3.3",
    "drf-spectacular==0.27.2",
    "idna==2.8",
    "itypes==1.1.0",
    "markupsafe==1.1.1",
    "openapi-codec==1.3.2",
    "pillow>=11.3.0",
    "psycopg>=3.1.8",
    "ptvsd==4.3.2",
    "pyjwt==2.5.0",
    "pylint==3.3.8",
    "pytz==2019.1",
    "requests==2.32.4",
    "simplejson==3.16.0",
    "sqlparse>=0.3.1",
    "uritemplate==3.0.0",
    "urllib3==2.2.0",
]
